pyproject.toml
[tool.pdm]


[project]
name = "ghpolicy"
version = "0.0.dev"
description = ""
authors = [
    {name = "Lab Digital", email = "opensource@labdigital.nl"},
]
dependencies = [
    "PyGithub>=1.58.1",
    "click>=8.1.3",
    "PyYAML>=6.0",
]
requires-python = ">=3.11"
license = {text = "Private"}

[build-system]
requires = ["pdm-pep517>=1.0.0"]
build-backend = "pdm.pep517.api"

[project.scripts]
ghpolicy = "ghpolicy.cli:invoke"
